+++ pkgs/code_builder/test/specs/code/expression_test.dart
@@ -867,7 +867,7 @@ void main() {
         equalsDart('late String foo = bar'));
   });
 
-  test('should emit a perenthesized epression', () {
+  test('should emit a parenthesized expression', () {
     expect(
         refer('foo').ifNullThen(refer('FormatException')
             .newInstance([literalString('missing foo')])
@@ -876,84 +876,84 @@ void main() {
         equalsDart('foo ?? (throw FormatException(\'missing foo\'))'));
   });
 
-  test('should emit an addition assigment expression', () {
+  test('should emit an addition assignment expression', () {
     expect(
       refer('foo').addAssign(refer('bar')),
       equalsDart('foo += bar'),
     );
   });
 
-  test('should emit a subtraction assigment expression', () {
+  test('should emit a subtraction assignment expression', () {
     expect(
       refer('foo').subtractAssign(refer('bar')),
       equalsDart('foo -= bar'),
     );
   });
 
-  test('should emit a multiplication assigment expression', () {
+  test('should emit a multiplication assignment expression', () {
     expect(
       refer('foo').multiplyAssign(refer('bar')),
       equalsDart('foo *= bar'),
     );
   });
 
-  test('should emit a division assigment expression', () {
+  test('should emit a division assignment expression', () {
     expect(
       refer('foo').divideAssign(refer('bar')),
       equalsDart('foo /= bar'),
     );
   });
 
-  test('should emit an int division assigment expression', () {
+  test('should emit an int division assignment expression', () {
     expect(
       refer('foo').intDivideAssign(refer('bar')),
       equalsDart('foo ~/= bar'),
     );
   });
 
-  test('should emit a euclidean modulo assigment expression', () {
+  test('should emit a euclidean modulo assignment expression', () {
     expect(
       refer('foo').euclideanModuloAssign(refer('bar')),
       equalsDart('foo %= bar'),
     );
   });
 
-  test('should emit a shift left assigment expression', () {
+  test('should emit a shift left assignment expression', () {
     expect(
       refer('foo').shiftLeftAssign(refer('bar')),
       equalsDart('foo <<= bar'),
     );
   });
 
-  test('should emit a shift right assigment expression', () {
+  test('should emit a shift right assignment expression', () {
     expect(
       refer('foo').shiftRightAssign(refer('bar')),
       equalsDart('foo >>= bar'),
     );
   });
 
-  test('should emit a shift right unsigned assigment expression', () {
+  test('should emit a shift right unsigned assignment expression', () {
     expect(
       refer('foo').shiftRightUnsignedAssign(refer('bar')),
       equalsDart('foo >>>= bar'),
     );
   });
 
-  test('should emit a bitwise AND assigment expression', () {
+  test('should emit a bitwise AND assignment expression', () {
     expect(
       refer('foo').bitwiseAndAssign(refer('bar')),
       equalsDart('foo &= bar'),
     );
   });
 
-  test('should emit a bitwise XOR assigment expression', () {
+  test('should emit a bitwise XOR assignment expression', () {
     expect(
       refer('foo').bitwiseXorAssign(refer('bar')),
       equalsDart('foo ^= bar'),
     );
   });
 
-  test('should emit a bitwise OR assigment expression', () {
+  test('should emit a bitwise OR assignment expression', () {
     expect(
       refer('foo').bitwiseOrAssign(refer('bar')),
       equalsDart('foo |= bar'),